Add methods to update entities in the openBIS instance
These methods can be used to update selected attributes
of individual entities in the openBIS instance.

All update objects require the id of an entity that will
be updated. Please note that some of the entity fields cannot
be changed once an entity is created, e.g. sample code
becomes immutable after creation.